+++ Configuration.Designer.cs
@@ -134,6 +134,7 @@
             this.MaximizeBox = false;
             this.MinimizeBox = false;
             this.Name = "Configuration";
+            this.ShowIcon = false;
             this.ShowInTaskbar = false;
             this.ResumeLayout(false);
             this.PerformLayout();