+++ Saut.Designer.cs
@@ -70,6 +70,7 @@
             this.MaximizeBox = false;
             this.MinimizeBox = false;
             this.Name = "Saut";
+            this.ShowIcon = false;
             this.ShowInTaskbar = false;
             this.ResumeLayout(false);
             this.PerformLayout();