+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-
+﻿
 ## Changelog du Gestionnaire de Fond d'Écran
 
 ### Version 0.8.0 (26 Février 2017) :
@@ -6,6 +6,7 @@
  - Corrections de quelques bugs.
  - Bouton pour réinitialiser les paramètres.
  - Suppression d'une fenêtre inutile "SelectionAlt" et utilisation plus intélligente de la fenêtre "Selection".
+ - Mise en place d'une fenêtre d'attente lors du chargement de la liste des fichiers.
 	
 ### Version 0.7.2 (24 Février 2017) :
  - Création du changelog.